requirements_infra.txt
aws-cdk-lib==2.139.0
constructs==10.1.93
streamlit
pandas
